fix a bug where deleted items would return
Deleted items were not saved to the db before. This would bring them
back in the following scenario with the clients A and B:

A and B start a collaborative session. They create elements. B leaves.
A deletes an item. A leaves. B returns. B would now see the deleted item
as the reconciliation algo has no chance to notice that the item which
is in B's localStorage but is missing in the room that the server sent
was in fact deleted. If B now makes a full sync, the Items get added
back on the server.